ConstellationPredictor/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
django==5.2.3
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gtts==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.4
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
numba==0.61.2
numpy==2.2.6
openai-whisper==20231117
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt-toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pygame==2.6.1
pyjwt==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
pyyaml==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
telegram==0.0.1
tiktoken==0.9.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.160
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
